requirement.txt
flask==2.2.5
tensorflow==2.12.0
opencv-python-headless==4.8.0.74
numpy==1.24.3
matplotlib==3.7.1
pillow==9.5.0
scikit-image==0.21.0